requirements.txt
requests
psycopg2-binary
amazon-paapi5